+++ main/fopen_wrappers.c
@@ -17,7 +17,7 @@
    +----------------------------------------------------------------------+
  */
 
-/* $Id: fopen_wrappers.c,v 1.166 2003/07/30 17:55:06 iliaa Exp $ */
+/* $Id: fopen_wrappers.c,v 1.167 2003/09/26 08:09:52 hholzgra Exp $ */
 
 /* {{{ includes
  */
@@ -299,7 +299,7 @@ PHPAPI int php_fopen_primary_script(zend_file_handle *file_handle TSRMLS_DC)
 		if (s) {			/* if there is no path name after the file, do not bother */
 							/* to try open the directory */
 			length = s - (path_info + 2);
-			if (length > sizeof(user) - 1)
+			if (length > (int)sizeof(user) - 1)
 				length = sizeof(user) - 1;
 			memcpy(user, path_info + 2, length);
 			user[length] = '\0';
